+++ rust/uefi/stub/src/common.rs
@@ -6,6 +6,7 @@ use uefi::{
 };
 
 use linux_bootloader::linux_loader::InitrdLoader;
+use linux_bootloader::measure::measure_boot_loader;
 use linux_bootloader::pe_loader::Image;
 use linux_bootloader::pe_section::pe_section_as_string;
 
@@ -20,28 +21,46 @@ pub fn extract_string(pe_data: &[u8], section: &str) -> uefi::Result<CString16> 
 ///
 /// If Secure Boot is active, this is always the embedded one (since the one passed from the bootloader may come from a malicious type 1 entry).
 /// If Secure Boot is not active, the command line passed from the bootloader is used, falling back to the embedded one.
-pub fn get_cmdline(embedded: &CStr16, secure_boot_enabled: bool) -> Vec<u8> {
-    if secure_boot_enabled {
-        // The command line passed from the bootloader cannot be trusted, so it is not used when Secure Boot is active.
-        embedded.as_bytes().to_vec()
-    } else {
-        let passed = boot::open_protocol_exclusive::<LoadedImage>(boot::image_handle())
-            .map(|loaded_image| loaded_image.load_options_as_bytes().map(|b| b.to_vec()));
-        match passed {
-            Ok(Some(passed)) => passed,
-            // If anything went wrong, fall back to the embedded command line.
-            _ => embedded.as_bytes().to_vec(),
+///
+/// If we read a user provided cmdline, we measure it into PCR4 to invalidate the PCR. This way we
+/// can support Measured Boot without Secure Boot while still giving users the option to edit the
+/// kernel cmdline on the fly for debugging.
+pub fn get_cmdline(embedded: &CStr16) -> Vec<u8> {
+    let secure_boot_enabled = get_secure_boot_status();
+    if !secure_boot_enabled {
+        let load_options = get_load_options();
+        if let Some(passed) = load_options {
+            // The passed cmdline is allowed to be the same as the embedded one without needing to
+            // be measured.
+            if passed == embedded.as_bytes() {
+                return passed;
+            }
+            // Measuring is mandatory here. If the measurement fails, return the embedded cmdline.
+            if measure_boot_loader(&passed, "Custom user provided kernel cmdline").is_ok() {
+                return passed;
+            }
         }
     }
+    embedded.as_bytes().to_vec()
 }
 
-/// Check whether Secure Boot is active, and we should be enforcing integrity checks.
+/// Obtain the load options of the currently loaded image.
+///
+/// This is used for example to pass a customer user provided kernel cmdline from the boot loader
+/// to the booted image.
+fn get_load_options() -> Option<Vec<u8>> {
+    boot::open_protocol_exclusive::<LoadedImage>(boot::image_handle())
+        .ok()
+        .and_then(|loaded_image| loaded_image.load_options_as_bytes().map(|b| b.to_vec()))
+}
+
+// Check whether Secure Boot is active.
 ///
 /// In case of doubt, true is returned to be on the safe side.
 pub fn get_secure_boot_status() -> bool {
     // The firmware initialized SecureBoot to 1 if performing signature checks, and 0 if it doesn't.
     // Applications are not supposed to modify this variable (in particular, don't change the value from 1 to 0).
-    let secure_boot_enabled = runtime::get_variable(
+    runtime::get_variable(
         cstr16!("SecureBoot"),
         &VariableVendor(guid!("8be4df61-93ca-11d2-aa0d-00e098032b8c")),
         &mut [1],
@@ -64,13 +83,7 @@ pub fn get_secure_boot_status() -> bool {
             warn!("Failed to read SecureBoot variable: {err}. Performing verification anyway.");
             true
         }
-    });
-
-    if !secure_boot_enabled {
-        warn!("Secure Boot is not active!");
-    }
-
-    secure_boot_enabled
+    })
 }
 
 /// Boot the Linux kernel without checking the PE signature.